PCI-SRIOV provider network shows as Down in red
The provider networks using PCI-PT and SRIOV interfaces are shown as
DOWN in red.  This indicates that there is no dhcp and virtual router
services for the networks attached to the provider networks configured
against the PCI-PT and SRIOV interfaces.

This update is to introduce a new status to represent this state
instead of showing it as Down.

Story: 2003408
Task: 24549

Change-Id: Ib9d22f7e5dd662c0456da28b85ed495568242f2a
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>