No need for PCI interface class options during interface creation
A change has been made in forms.py to remove
'pci-passthrough' and 'pci-sriov' as an ifclass when creating an
interface. As the code stands, provider networks are only listed for
either 'pci-passthrough' and 'pci-sriov' when editing an interface, and
a PCI interface cannot be created if a provider network is not selected.
Thus, those options have been removed from the creation form.

Closes-Bug: 1795717

Change-Id: I898d21807d16375d3ccaa40c29a3920656aad44b
Signed-off-by: Patrick Bonnell <patrick.bonnell@windriver.com>