Use timesinceSorter instead of the uptimeSorter.
timesinceSorter is provided by horizon as the default sorter of
the datetime fields in a table. We don't have to create a uptimeSorter
for host inventory panel.

The reason that we failed to use timesinceSorter before is the
datetime returned by host.boottime is calculated by datetime.now() but
not timezone.now(). With this change, timesinceSorter can work well
with host inventory panel.

Test done for the change:

1. Build iso with this patch.
2. Deploy with 1 controller and 2 compute nodes.
3. Open starlingx dashboard, Admin -> Platform -> Host Inventory,
press on the column head of "Uptime" of the table for compute nodes,
the timesinceSorter works well to sort the compute nodes.
4. Check the page source, the Uptime column is marked as
"data-type="timesince"", and the data for this column is marked
with data-seconds (added by the timesince_sortable filter),
format as below:
  <span data-seconds="4361">1 hour, 12 minutes</span>

Closes-Bug: 1809861

Change-Id: I5f9e896ccb0e4bb0e7d801402f9139e191cf9e08
Signed-off-by: Yan Chen <yan.chen@intel.com>