Enable configurable vswitch memory
Currently, a DPDK enabled vswitch makes use of a fixed 1G hugepage to
enable an optimized datapath.

In the case of OVS-DPDK, this can cause an issue when changing the
MTU of one or more interfaces, as a seperate mempool is allocated
for each size.  If the minimal mempool size(s) cannot fit into the
1G page, DPDK memory initialization will fail.

This commit allows an operator to configure the amount of hugepage
memory allocated to each socket on a host, which can enable
jumboframe support for OVS-DPDK.

Change-Id: Ie54cc23a0839d80d2062510f8f6128eb68b2cfc9
Story: 2004472
Task: 28162
Depends-On: https://review.openstack.org/#/c/631039
Signed-off-by: Steven Webster <steven.webster@windriver.com>