Refactor the FM panel to leverage AngularJS
The FM panel which contains 3 tabs is now replaced with a new panel
group called “Fault Management” under Admin which contains an
“Active Alarms”, “Events” and “Events Suppression” panel.

The suppression buttons are gone, filters replace the need for the
suppression buttons. Paging is done automatically in AngularJS the and
‘Default Limit' button is also gone.

The Alarm Summary Banner was removed from the FM panels as it is
already displayed in the header banner.

The “Related Alarms” tab under the “Provider Network Topology“ panel was
refactored to still uses Django’s alarms table.

Other related FM Django code that is no longer used has been removed.

Refactored some DC Cloud panels as well.

Story: 2004818
Task: 28984

Change-Id: I7fdece26118dc066cf93001189767f78bfe2caf7
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>