Create DataNetworks modelling in System Configuration
Introduce the stx-gui updates to enable DataNetwork modelling of the
physical data network in sysinv.  This is part of the Story to
"Move neutron provider network modelling to system configuration".

This is an initial update to enable Interfaces, Data Networks and
Data Network Topology panels.

Pending Stein spec support for network-segment-range-mangement,
the following GUI features are not currently supported:
    Segmentation Ranges
    Tenant Networks

The following GUI panels are updated:
    Host Inventory - Interfaces
    Data Networks (<oamip> /admin/datanets; previously /admin/providernets)
    Data Network Topology

Tests Performed:
    Data Network create, modify for network_type flat, vlan.
        vxlan excluded in this GUI commit.
    Data Network delete, modify
    Data Network Topology - basic navigation to panel
    Interface Profile create and apply
    Sanity

Story: 2004455
Task: 28325

Depends-On: https://review.openstack.org/#/c/631701/
Change-Id: I46db5b5d47a9920f21db0521a8f5fbe4213b8af3
Signed-off-by: John Kung <john.kung@windriver.com>