Enhance drawer for alarms and events panel
Enhance the drawer for alarms and events panel to display the entire
data set and not just a subset. Remove the link to the details page
for alarms and events.

Remove the details page for events. The details page for alarms is
being used by the “Admin/Platform/Provider Network Topology“ panel.

Story: 2004818
Task: 29048

Change-Id: I81ac78416ab52ec2b8942be08ed3e5e12976a97d
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>