Fix horizon VM launch
Moving the cookie customization to the platform horizon, and
manually setting the new cookie to be used by our angularJS
FM client.  The change is not needed by the sysinv client since we are
able to embed the csrf token in the django template, which isn't
possible with the pure angular FM panels.

Change-Id: Iceebf7028325256b6793deb296d32e9a9f5fba21
Signed-off-by: Tyler Smith <tyler.smith@windriver.com>
Closes-Bug: 1813661