Set the ApiAccess horizon page as the default on login.
The previous default was a nova overview page, which caused
horizon to fail with too many redirects if the nova endpoints
are not setup.

This change allows the nova endpoints and puppet manifests
to be skipped when setting up the controller.

This change also sets the revision to be based on git commits.

Story: 2004764
Task: 29839
Change-Id: Ice10fe6da6b34f1d9206f26e112eb555e2088932
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>